+++ doc/changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+# 1.1.1 - 2017-05-04
+### Added
+* Data Retreiver Helper
+* getLocalData function
+* getRemoteData function
+* downloadFile function
+* listFiles function
+
 # 1.1.0 - 2017-03-14
 ### Added
 * Database class can now import from compressed archive.